backend/requirements.txt
google-generativeai
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
python-dotenv==1.0.0
flask-cors==4.0.0
Pillow==10.1.0
openai
reportlab==4.0.7
anthropic==0.40.0
requests==2.31.0
beautifulsoup4==4.12.2
opencv-python==4.8.1.78
mediapipe==0.10.21
numpy==1.26.4
pytesseract
pdfplumber
pyhwp

